Added RdsSharedDBPool and RdsDbConnection
In order to help generalize the types of RDS (postgres, mysql, etc) and
to support multiple shared plans for the different types, the
RdsSharedDbPool is created.

The pool is just a mapping of the shared plan id (from the catalog.go)
to the RdsDbConnection which contains the persistent connection to the
database.

The broker still uses a database to maintain record of its actions and
states. However, this patch now removes the assumption that it will be
the only connected shared DB. The broker can now choose which shared DB
it wants to use by specifying the plan id of the database it should use
by setting the BROKER_DB_PLAN_ID environment variable.