Added DBAdapterFactoryInterface
Adding the general interface for the DBAdpaterInterface so that it can
specific implementations can control the types of SharedDB and
DedicatedDB adapters created. This is useful for testing as we can
create mock factories that can create adapters that can logically
simulate DB interactions. This was needed as some syntax used in MySQL
and Postgres is not supported by SQLITE. Therefore, tests would fail
once we tried to do things like CREATE DATABASE. It would work in
Postgres in production but not SQLITE in unit tests.