Added CreateDBAdapter and DeleteDB methods.
DeleteDB is added to the interface of DBAdapter so that the controller
code does not have to worry about what type of DBAdapter (shared v.
dedicated) is being used and only needs to worry about the returned
status (and error if it exists).

CreateDBAdapter replaces the CreateDB method for the DBAdapterFactory.
The original CreateDB for the DBAdapterFactory took care of figuring
out which adapter to use and then calling CreateDB with the appropriate
adapter. With the addition of more methods that we want to use with the
adapter, instead of copying a lot of the logic with the addition of a
new adapter method, we just instead return the adapter itself and the
caller can then decide what method (new or existing; in our case the
DeleteDB) to call.

TODO: Implement the DeleteDB for the dedicated instance.