Added BindDBToApp to DBAdapter and GetCredentials to Instance
BindDBToApp allows each individual DBAdapter to take care of how they
bind dbs to apps. This is useful when comparing binding a SharedDB
versus a DedicatedDB. The SharedDB will be ready instantly whereas the
DedicatedDB may still be creating. The BindDBToApp for DedicatedDB
adapter will appropriately query for instance state and only
successfully bind to an app that is ready.

GetCredentials separate the logic so that we can generate the
credentials appropriately on a per-dbtype basis (postgres, etc).